Heart Disease Prediction System
In medical field there is a growing need for development of software to be used for practical applications rather than only in research field. Doctors need not only to know the main symptoms in a patient but also the other factors in life and surrounding that lead to disease in the patient. For collecting all the necessary information about the patient the doctors needs to spend a lot of time with the patient asking necessary questions to get the important details. The doctor needs to document all the data for further use and needs to keep track of all the patient details and remedies. The proposed system is intended to develop an Intelligent System using data mining modeling technique, namely, Naive Bayes. It is implemented as Java application in which user answers the predefined questions. It retrieves hidden data from stored database and compares the user values with trained data set. It can answer complex queries for diagnosing heart disease and thus assist healthcare practitioners to make intelligent clinical decisions which traditional decision support systems cannot.
Consulted Medical Practitioner to understand symptoms of Heart Diseases and Remedies. Published Research Paper at IEEE Conference.
Paper Link - ‘http://ieeexplore.ieee.org/document/7095854/?reload=true&arnumber=7095854&searchWithin=%22Authors%22:.QT.S.’